Added vehicle service layer, added admin server vehicle feature, added login and admin page security filter, improved RentalServlet code

+++ src/main/java/app/services/admin/DefaultAdminService.java
@@ -1,14 +1,22 @@
 package app.services.admin;
 
+import java.util.Arrays;
+import java.util.Date;
+import java.util.List;
+
+import app.dao.reservation.ReservationDao;
+import app.model.Reservation;
 import app.model.Vehicle;
 import app.services.vehicle.VehicleService;
 
 public class DefaultAdminService implements AdminService {
 	
 	private VehicleService vehicleService;
+	private ReservationDao reservationDao;
 
-	public DefaultAdminService(VehicleService vehicleService) {
+	public DefaultAdminService(VehicleService vehicleService, ReservationDao reservationDao) {
 		this.vehicleService = vehicleService;
+		this.reservationDao = reservationDao;
 	}
 
 	@Override
@@ -18,12 +26,26 @@ public class DefaultAdminService implements AdminService {
 
 	@Override
 	public void updateVehicle(Vehicle vehicle) {
-		vehicleService.update(vehicle);
+		if (!this.vehicleHasFutureReservation(vehicle.getId())) {
+			vehicleService.update(vehicle);
+		} else {
+			throw new IllegalArgumentException("Cannot update a vehicle with a reservation in the future");
+		}
 	}
 
 	@Override
 	public void deleteVehicle(long vehicleId) {
-		vehicleService.delete(vehicleId);
+		if (!this.vehicleHasFutureReservation(vehicleId)) {
+			vehicleService.delete(vehicleId);
+		} else {
+			throw new IllegalArgumentException("Cannot delete a vehicle with a reservation in the future");
+		}
+	}
+	
+	private boolean vehicleHasFutureReservation(long vehicleId) {
+		List<Reservation> reservations = reservationDao.findByVehicleIds(Arrays.asList(vehicleId));
+		Date today = new Date();
+		return reservations.stream().anyMatch(reservation -> reservation.getDropoffDate().after(today));
 	}
 
 }
